refactor components and deps.
