+++ libs/design-system/src/lib/Components/AnnouncementCard/index.tsx
@@ -1,4 +1,4 @@
-import React from 'react';
+
 import { formatDistance } from 'date-fns';
 import { ReactComponent as ChevronRightIcon } from '../../../../images/arrows/chevron-right.svg';
 import { ReactComponent as TimeIcon } from '../../../../images/content/date.svg';
@@ -21,9 +21,9 @@ const AnnouncementCard = ({
   return (
     <a
       href={link}
-      className="flex items-center justify-between rounded-2xl bg-white px-4 py-9 hover:shadow-2xl dark:bg-primary-gray-dark md:px-8 md:py-6"
+      className="flex items-center justify-between px-4 bg-white rounded-2xl py-9 hover:shadow-2xl dark:bg-primary-gray-dark md:px-8 md:py-6"
     >
-      <div className="mr-1 mr-4 self-start">
+      <div className="self-start mr-1 mr-4">
         <img
           src={sourceIcon}
           alt={sourceAltText}
@@ -34,7 +34,7 @@ const AnnouncementCard = ({
       </div>
       <div className="flex-1">
         <div className="text-xl text-2xl font-bold">{heading}</div>
-        <div className="mt-4 flex items-center text-primary-gray-300">
+        <div className="flex items-center mt-4 text-primary-gray-300">
           <TimeIcon />
           <span className="ml-2">
             {formatDistance(timestamp, new Date())} ago
