+++ libs/design-system/src/lib/Components/Callout/index.tsx
@@ -1,5 +1,3 @@
-import React from 'react';
-
 export type CalloutProps = {
   heading: string;
   description: string;
@@ -10,7 +8,7 @@ export type CalloutProps = {
 const Callout = ({ heading, description, ctaText, ctaLink }: CalloutProps) => {
   return (
     <div
-      className="flex flex-col items-start justify-between rounded-2xl p-10 dark:bg-primary-gray-dark md:flex-row md:items-center md:px-20 md:pt-16 md:pb-12"
+      className="flex flex-col items-start justify-between p-10 rounded-2xl dark:bg-primary-gray-dark md:flex-row md:items-center md:px-20 md:pt-16 md:pb-12"
       style={{ backgroundColor: 'rgba(222,226,233, 0.5)' }}
     >
       <div className="flex-1">
@@ -18,7 +16,7 @@ const Callout = ({ heading, description, ctaText, ctaLink }: CalloutProps) => {
         <p className="pt-2 dark:text-primary-gray-100">{description}</p>
       </div>
       <a
-        className="h-14 rounded-lg bg-black px-16 py-5 text-sm text-white hover:cursor-pointer"
+        className="px-16 py-5 text-sm text-white bg-black rounded-lg h-14 hover:cursor-pointer"
         href={ctaLink}
       >
         {ctaText}
