+++ libs/design-system/src/lib/Components/CommunityMembers/index.tsx
@@ -1,4 +1,3 @@
-import React from 'react';
 import { RoundImage } from '../';
 
 export type CommunityMembersProps = {
