+++ libs/design-system/src/lib/Components/FeaturedArticleCard/index.tsx
@@ -1,4 +1,4 @@
-import React from 'react';
+ 
 import { ButtonLink } from '../Button';
 import Tag from '../Tag';
