+++ libs/design-system/src/lib/Components/FlipCell/index.tsx
@@ -1,4 +1,4 @@
-import React from 'react';
+ 
 import { ReactComponent as CalendarIcon } from '../../../../images/action/date-calendar.svg';
 import { ReactComponent as CommentIcon } from '../../../../images/arrows/message-circle.svg';
 import Tag from '../Tag';
