+++ libs/design-system/src/lib/Components/Flips/RoundImage.tsx
@@ -1,4 +1,4 @@
-import React from 'react';
+ 
 
 export type RoundImageProps = {
   imageUri: string;
