+++ libs/design-system/src/lib/Components/LandingHeader/LandingHeaderContainer.tsx
@@ -1,5 +1,5 @@
 import clsx from 'clsx';
-import React from 'react';
+ 
 
 export type GradientName =
   | 'community'
