+++ libs/design-system/src/lib/Components/RoundImage/index.tsx
@@ -1,4 +1,4 @@
-import React from 'react';
+ 
 import clsx from 'clsx';
 
 export type RoundImageProps = {
