+++ libs/design-system/src/lib/Components/SocialLinksSignup/index.tsx
@@ -1,4 +1,4 @@
-import React from 'react';
+ 
 import { ReactComponent as GithubLogo } from '../../../../images/social/github.svg';
 import { ReactComponent as DiscordLogo } from '../../../../images/social/discord.svg';
 import { ReactComponent as DiscourseLogo } from '../../../../images/social/forum.svg';
