+++ libs/design-system/src/lib/Components/SporksCard/index.tsx
@@ -1,5 +1,5 @@
 import clsx from 'clsx';
-import React, { useState } from 'react';
+import   { useState } from 'react';
 import { format, formatDistanceToNow } from 'date-fns';
 import { ReactComponent as CopyIcon } from '../../../../images/action/copy.svg';
 import { ReactComponent as ChevronUpIcon } from '../../../../images/arrows/chevron-up.svg';
