+++ libs/design-system/src/lib/Pages/NetworkPage/index.tsx
@@ -1,4 +1,4 @@
-import React from 'react';
+ 
 import { startOfDay } from 'date-fns';
 import {
   NetworkCard,
