+++ nx.json
@@ -5,7 +5,8 @@
     "defaultBase": "main"
   },
   "cli": {
-    "defaultCollection": "@nrwl/react"
+    "defaultCollection": "@nrwl/react",
+    "packageManager": "yarn"
   },
   "implicitDependencies": {
     "package.json": {
