+++ package.json
@@ -69,6 +69,7 @@
     "eslint-plugin-react": "7.29.4",
     "eslint-plugin-react-hooks": "4.4.0",
     "jest": "27.2.3",
+    "lerna": "^5.0.0",
     "nx": "14.1.6",
     "postcss": "^8.4.13",
     "postcss-import": "^14.1.0",
@@ -97,8 +98,8 @@
     "@remix-run/serve": "^1.0.6",
     "core-js": "^3.6.5",
     "npm-run-all": "^4.1.5",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "react": "^18.0.0",
+    "react-dom": "^18.0.0",
     "regenerator-runtime": "0.13.7",
     "remix": "^1.0.6",
     "tailwindcss": "^3.0.24",
